requirements.txt
numpy
pytube
sanic
humps
scipy
spacy

### old, not compatible with spacy3.7.1
# https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz

### already installed... pip wont install again, but first downloads tar file... so, commenting
# https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
# https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1.tar.gz

gensim
scikit-learn

pandas
vaderSentiment
fasttext

dill        ### 

aiohttp
asyncio
async_timeout

boto3
pyenchant